OOP makeover
Classes to define a Mini-framework that will allow us to do things much
easier in the future. Plus Classes corresponding to the tables in our
database so that we can take advantage of PDO database access.

Also signup and signing functionality is complete, except for “remember
me” which I will add later.