A lot more stuff
Home page is basically done, so that’s more stuff you can look at for
reference.

Also, I filled out the Track class to include a method to display its
mini version. Due to the OOP you can just call that function on a track
object and it handles everything. Also I used the include functionality
so that the html for it is in a separate file.