Almost Done
Added Search Capabilities, and the news feed.
Also got the liking/unliking of tracks on home to work (liking one
track was effecting all others on the page before). I will make the
changes to fix the like/unlike on the track page and for replies as
well tomorrow and then its pretty much done.